change database provider

+++ DictionaryBot/EventHandlers/DictionaryEventHandler.cs
@@ -9,10 +9,10 @@ namespace DictionaryBot.EventHandlers
 {
     internal class DictionaryEventHandler
     {
-        private static string _lastWordCache = "";
-        private static ulong _lastUserCache = 0;
+        private static readonly Dictionary<ulong, string> _lastWordCache = [];
+        private static readonly Dictionary<ulong, ulong> _lastUserCache = [];
 
-        internal static Task GuildAvailable(DiscordClient sender, GuildCreateEventArgs e)
+        internal static Task GuildAvailable(DiscordClient _1, GuildCreatedEventArgs e)
         {
             _ = Task.Run(async () =>
             {
@@ -33,16 +33,18 @@ namespace DictionaryBot.EventHandlers
                     return;
 
                 //cache last (hopefully) valid message
-                var channel = e.Guild.GetChannel((ulong)channelId);
-                var messages = await channel.GetMessagesAsync(20);
-                var message = messages.First(x => !x.Content.Trim().StartsWith("."));
-                _lastWordCache = message.Content.Trim();
-                _lastUserCache = message.Author.Id;
+                var channel = await e.Guild.GetChannelAsync((ulong)channelId);
+                var messages = channel.GetMessagesAsync(20);
+                var message = await messages.FirstAsync(x => !x.Content.Trim().StartsWith('.'));
+                _lastWordCache[e.Guild.Id] = message.Content.Trim();
+                if (message.Author is null)
+                    return;
+                _lastUserCache[e.Guild.Id] = message.Author.Id;
             });
             return Task.CompletedTask;
         }
 
-        internal static Task GuildCreated(DiscordClient sender, GuildCreateEventArgs e)
+        internal static Task GuildCreated(DiscordClient _1, GuildCreatedEventArgs e)
         {
             _ = Task.Run(async () =>
             {
@@ -63,29 +65,31 @@ namespace DictionaryBot.EventHandlers
                     return;
 
                 //cache last (hopefully) valid message
-                var channel = e.Guild.GetChannel((ulong)channelId);
-                var messages = await channel.GetMessagesAsync(20);
-                var message = messages.First(x => !x.Content.Trim().StartsWith("."));
-                _lastWordCache = message.Content.Trim();
-                _lastUserCache = message.Author.Id;
+                var channel = await e.Guild.GetChannelAsync((ulong)channelId);
+                var messages = channel.GetMessagesAsync(20);
+                var message = await messages.FirstAsync(x => !x.Content.Trim().StartsWith('.'));
+                _lastWordCache[e.Guild.Id] = message.Content.Trim();
+                if (message.Author is null)
+                    return;
+                _lastUserCache[e.Guild.Id] = message.Author.Id;
             });
             return Task.CompletedTask;
         }
 
-        internal static Task MessageCreated(DiscordClient _1, MessageCreateEventArgs e)
+        internal static Task MessageCreated(DiscordClient _1, MessageCreatedEventArgs e)
         {
             _ = Task.Run(async () =>
             {
                 if (e.Author.IsBot)
                     return;
-                if (e.Message.Content.StartsWith("."))
+                if (e.Message.Content.StartsWith('.'))
                     return;
                 using DatabaseContext db = new();
                 if (e.Channel.Id != db.Guilds.Find(e.Guild.Id)?.DictionaryGameChannel)
                     return;
 
 
-                if (e.Message.Author.Id == _lastUserCache)
+                if (e.Message.Author is not null && e.Message.Author.Id == _lastUserCache[e.Guild.Id])
                 {
                     await e.Message.DeleteAsync(); //delete message
                     var msg = await e.Channel.SendMessageAsync($"{e.Author.Mention} wait for someone else to send a message!"); //inform user that he has to wait 
@@ -94,7 +98,7 @@ namespace DictionaryBot.EventHandlers
                     return;
                 }
 
-                if (e.Message.Content.Trim().IndexOf(' ') != -1) //check for spaces in the message, if there are none move on
+                if (e.Message.Content.Trim().Contains(' ')) //check for spaces in the message, if there are none move on
                 {
                     await e.Message.DeleteAsync(); //delete message
                     var msg = await e.Channel.SendMessageAsync($"{e.Author.Mention} your message cannot have spaces!"); //inform user message cant have whitespaces
@@ -103,9 +107,9 @@ namespace DictionaryBot.EventHandlers
                     return;
                 }
 
-                if (!string.IsNullOrWhiteSpace(_lastWordCache))
+                if (!string.IsNullOrWhiteSpace(_lastWordCache[e.Guild.Id]))
                 {
-                    if (!e.Message.Content.Trim().ToLower().StartsWith(_lastWordCache.ToLower().Last()))
+                    if (!e.Message.Content.Trim().ToLower().StartsWith(_lastWordCache[e.Guild.Id].ToLower().Last()))
                     {
                         await e.Message.DeleteAsync(); //delete message
                         var msg = await e.Channel.SendMessageAsync($"{e.Author.Mention} the word {e.Message.Content.Trim()} does not start with an {_lastWordCache.Last()}!"); //inform user that chars have to match
@@ -133,14 +137,15 @@ namespace DictionaryBot.EventHandlers
                     await msg.DeleteAsync(); //delete information again
                     return;
                 }
-
                 await e.Message.CreateReactionAsync(DiscordEmoji.FromUnicode("✅")); //add reaction
 
                 db.Add(new DictionaryEntry(e.Message.Content.Trim()) { GuildId = e.Guild.Id }); //populate db
                 db.SaveChanges();
 
-                _lastWordCache = e.Message.Content.Trim(); //populate cache
-                _lastUserCache = e.Message.Author.Id;
+                _lastWordCache[e.Guild.Id] = e.Message.Content.Trim(); //populate cache
+                if (e.Message.Author is null)
+                    return;
+                _lastUserCache[e.Guild.Id] = e.Message.Author.Id;
             });
             return Task.CompletedTask;
         }
